api/requirements.txt
libgravatar==1.0.0
djangorestframework_simplejwt==4.8.0
Django==2.2.12
djangorestframework==3.12.4
psycopg2-binary==2.8.6
django-cors-headers==3.8.0
cloudinary==1.26.0 
python-dotenv==0.19.0 